[RENAME] improves folder structure
