+++ src/router/index.js
@@ -21,7 +21,9 @@ import createDashboard from "@/components/Dashboards/createDashboard"
 import editDashboard from "@/components/Dashboards/editDashboard"
 
 // tags
-import CreateTag from "@/components/Projects/CreateTag"
+import CreateTag from "@/components/Tags/CreateTag"
+
+// Router guard
 import Guard from "@/components/Auth/middleware"
 
 Vue.use(Router)
